Successfully parse 'a + b /= c'
We attempt to parse '/=' before we parse '/', which parses successfully.
But because in this expression we previously encountered a '+', we're at a precedence leve where we are only evaluating operators with a higher precedence than '/='. So we don't check for '/=', and then we see a '/' which commits, and then we get stuck because we don't expect a '='.

This fixes the issue by checking, for the '/' operator, that it is not
followed by a '='. Ideally we should only do this if the operator
precedence is higher than the precendence for '/=', but that data is not
available for the parser.

Similarly, this fixes the parsing of 'a && b <| c' which has the same
problem.